posting to database functional
