+++ businessServer/client/actions/actions.js
@@ -1,106 +1,3 @@
-import store from '../store'
-import fetch from 'isomorphic-fetch'
-
-
-export function saveNodeAction(nodeObj){
-    if(!nodeObj.question){
-        nodeObj.question = "default value";
-    }
-    store.dispatch(postingNode());
-    return fetch('/api/nodes/', {
-        method: 'POST',
-        header: {
-            'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({
-            question: nodeObj.question,
-            productId: nodeObj.productId
-        })
-    })
-    .then(res => res.json())
-    .then(item => {
-        console.log(item);
-        if(!item){
-            dispatch(errorPostingNode())
-        }
-        else{
-            dispatch(successPostingNode(item))
-        }
-    })
-
-}
-export function saveConnAction(connObj){
-    store.dispatch(postingConn());
-    return fetch('/api/connections/', {
-        method: 'POST',
-        header: {
-            'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({
-            answer: nodeObj.question,
-            fromId: nodeObj.fromId,
-            productId: connObj.productId,
-            businessId: connObj.businessId
-        })
-    })
-    .then(res => res.json())
-    .then(item => {
-        console.log(item);
-        if(!item){
-            dispatch(errorPostingConn())
-        }
-        else{
-            dispatch(successPostingConn(item))
-        }
-    })
-
-}
-
-export function postingNode(){
-    return {
-        type: "POSTING_NODE",
-        posting: true
-    }
-}
-
-export function errorPostingNode(){
-    return {
-        type: "POSTING_NODE_ERROR",
-        posting: false,
-        item: null
-    }
-}
-
-export function successPostingNode(item){
-    return {
-        type: "POSTING_NODE_SUCCESS",
-        posting: false,
-        item: item
-    }
-}
-
-export function postingConn(){
-    return {
-        type: "POSTING_CONN",
-        posting: true
-    }
-}
-
-export function errorPostingConn(){
-    return {
-        type: "POSTING_CONN_ERROR",
-        posting: false,
-        item: null
-    }
-}
-
-export function successPostingConn(item){
-    return {
-        type: "POSTING_CONN_SUCCESS",
-        posting: false,
-        item: item
-    }
-}
 
 export function addTopLayerNodeAction(newNodeId, productId, layer){
   return {
